Refactor:  - implemented AnimationExt.kt  - extracts to Constants.kt Changed use_controller to true to automatically hide them after sometime.
