+++ app/src/main/java/com/overplay/blackmagicshinedemo/constants/Constants.kt
@@ -3,5 +3,6 @@ package com.overplay.blackmagicshinedemo.constants
 class Constants {
     companion object {
         const val VIDEO_URL = "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/WeAreGoingOnBullrun.mp4"
+        const val DELAY_IN_SECONDS = 4
     }
 }
