+++ app/src/main/res/layout/activity_main.xml
@@ -12,7 +12,7 @@
         android:id="@+id/video_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:use_controller="false"/>
+        app:use_controller="true"/>
 
     <TextView
         android:id="@+id/countdown"
